Added to Chapter 3
Added a section after the introductory material describing how subcolumn cloud
and precipitation is generated within COSP. This is partly cut from the
Framework section below, and partly new material describing how SCOPS and
PREC_SCOPS are implemented in COSP. The idea is to completely describe the
treatment of subcolumn clouds and precipitation in this section. Most of the
discussion of SCOPS and PREC_SCOPS are done, but I still need to add discussion
of how condensate amount is then assigned to cloudy/precipitating columns.