Edits to the subgrid chapter (Chapter 3)
Continued cleaning up the Generation of subgrid profiles and the Framework
sections, and deleted redundant material between the introduction and these
sections. Started in on the results section, but have a ways to go. Removed the
overlap figure from Chapter 3, because I do not think it made as much sense
without the detailed description of how overlap is calculated that is provided
in Chapter 4. I think just a reference to the relevant literature showing that
the MRO overestimates vertical correlation of continuous cloud layers will be
sufficient.