Updated figures in chapter3.tex
Added the attenuation test figure in the first subgrid chapter, and updated the
caption in the placeholder. Added ISCCP cloud area by cloud top height maps as
well, and updated the text commenting on the similarities and differences with
the MISR figure. Updated the caption for the ISCCP figure as well. Made the
joint histogram figures, but opted not to include them here because they did not
add much to the discussion. I believe chapter3 is mostly complete now, pending
a re-read.