First version sent to advisers.
Last chapter is not complete yet. Had planned to make this basically a
re-pitching of the CMIP5 paper that was rejected, but with more intelligent
analyis focusing on the things we are confident about having done the work in
the previous chapters. Looking more like this will be just a conclusion though,
with commentary on relevance to model evaluation, and maybe some passing
references/figures incorporating the CMIP5 analysis into that.
`